+++ R/write.R
@@ -520,6 +520,7 @@ write_data2 <-function(wb, sheet, data, name = NULL,
 #' @param x A data frame.
 #' @param startCol A vector specifying the starting column to write df
 #' @param startRow A vector specifying the starting row to write df
+#' @param dims Spreadsheet dimensions that will determine startCol and startRow: "A1", "A1:B2", "A:B"
 #' @param array A bool if the function written is of type array
 #' @param xy An alternative to specifying startCol and startRow individually.
 #' A vector of the form c(startCol, startRow)
@@ -544,6 +545,7 @@ write_data_table <- function(
     x,
     startCol = 1,
     startRow = 1,
+    dims,
     array = FALSE,
     xy = NULL,
     colNames = TRUE,
@@ -577,6 +579,12 @@ write_data_table <- function(
   assert_class(bandedRows, "logical")
   assert_class(bandedCols, "logical")
 
+  if (!is.null(dims)) {
+    dims <- dims_to_rowcol(dims, as_integer = TRUE)
+    startCol <- min(dims[[1]])
+    startRow <- min(dims[[2]])
+  }
+
   if (missing(na.strings)) na.strings <- substitute()
 
   ## common part ---------------------------------------------------------------
@@ -591,6 +599,7 @@ write_data_table <- function(
 
   ## All input conversions/validations
   if (!is.null(xy)) {
+    warning("xy is deprecated. Please use dims instead.")
     if (length(xy) != 2) {
       stop("xy parameter must have length 2")
     }
@@ -786,6 +795,7 @@ write_data_table <- function(
 #' @param x Object to be written. For classes supported look at the examples.
 #' @param startCol A vector specifying the starting column to write to.
 #' @param startRow A vector specifying the starting row to write to.
+#' @param dims Spreadsheet dimensions that will determine startCol and startRow: "A1", "A1:B2", "A:B"
 #' @param array A bool if the function written is of type array
 #' @param xy An alternative to specifying `startCol` and
 #' `startRow` individually.  A vector of the form
@@ -870,6 +880,7 @@ write_data <- function(
     x,
     startCol = 1,
     startRow = 1,
+    dims = rowcol_to_dims(startRow, startCol),
     array = FALSE,
     xy = NULL,
     colNames = TRUE,
@@ -889,6 +900,7 @@ write_data <- function(
     x = x,
     startCol = startCol,
     startRow = startRow,
+    dims = dims,
     array = array,
     xy = xy,
     colNames = colNames,
@@ -927,6 +939,7 @@ write_data <- function(
 #' @param x A character vector.
 #' @param startCol A vector specifying the starting column to write to.
 #' @param startRow A vector specifying the starting row to write to.
+#' @param dims Spreadsheet dimensions that will determine startCol and startRow: "A1", "A1:B2", "A:B"
 #' @param array A bool if the function written is of type array
 #' @param xy An alternative to specifying `startCol` and
 #' `startRow` individually.  A vector of the form
@@ -1012,6 +1025,7 @@ write_formula <- function(wb,
   x,
   startCol = 1,
   startRow = 1,
+  dims = rowcol_to_dims(startRow, startCol),
   array = FALSE,
   xy = NULL) {
   assert_class(x, "character")
@@ -1028,6 +1042,7 @@ write_formula <- function(wb,
     x = dfx,
     startCol = startCol,
     startRow = startRow,
+    dims = dims,
     array = array,
     xy = xy,
     colNames = FALSE,
@@ -1044,6 +1059,7 @@ write_formula <- function(wb,
 #' @param x A data frame.
 #' @param startCol A vector specifying the starting column to write df
 #' @param startRow A vector specifying the starting row to write df
+#' @param dims Spreadsheet dimensions that will determine startCol and startRow: "A1", "A1:B2", "A:B"
 #' @param xy An alternative to specifying startCol and startRow individually.
 #' A vector of the form c(startCol, startRow)
 #' @param colNames If `TRUE`, column names of x are written.
@@ -1174,6 +1190,7 @@ write_datatable <- function(
     x,
     startCol = 1,
     startRow = 1,
+    dims = rowcol_to_dims(startRow, startCol),
     xy = NULL,
     colNames = TRUE,
     rowNames = FALSE,
@@ -1196,6 +1213,7 @@ write_datatable <- function(
     x = x,
     startCol = startCol,
     startRow = startRow,
+    dims = dims,
     array = FALSE,
     xy = xy,
     colNames = colNames,